Changed the Task Class and Added basic functionality to the Model and Storage components
The task class will still need another constructor that uses
CommandDetail as the argument.
Model does not support redo/undo functions yet.
For modfiy(), Logic will assume Parser knows what is changed and pass
the details of the changes as a commandDetail for creating a Task Object
which will have <null> fields for fields to remain unchanged and
non-null for otherwise. Note <null> != <emptyString>